atlas: response handling + authoritative section; todo additional section
