+++ plugin/atlas/atlas.go
@@ -5,14 +5,9 @@ package atlas
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	"github.com/coredns/coredns/plugin"
-	"github.com/coredns/coredns/plugin/atlas/ent"
-	"github.com/coredns/coredns/plugin/atlas/ent/dnsrr"
-	"github.com/coredns/coredns/plugin/atlas/ent/dnszone"
-	"github.com/coredns/coredns/plugin/atlas/record"
 	"github.com/coredns/coredns/plugin/metrics"
 	"github.com/coredns/coredns/request"
 
@@ -32,190 +27,113 @@ type Atlas struct {
 // in a Server.
 func (a *Atlas) ServeDNS(ctx context.Context, w dns.ResponseWriter, r *dns.Msg) (int, error) {
 
-	log.Infof("Atlas ServeDNS")
-	req := request.Request{W: w, Req: r}
-
-	reqName := req.Name()
-	reqType := req.QType()
-
-	client, err := a.getAtlasClient()
-	if err != nil {
-		return a.errorResponse(req, dns.RcodeServerFailure, err)
-	}
-	defer client.Close()
-
-	log.Info("question name: ", reqName)
-	log.Infof("question type: %v => %v", reqType, req.Type())
-
-	if time.Since(a.lastZoneUpdate) > a.cfg.zoneUpdateTime {
-		log.Info("++++++ LOADING ZONES +++++++")
-		err := a.loadZones(ctx, client)
-		if err != nil {
-			return a.errorResponse(req, dns.RcodeServerFailure, err)
-		}
-		a.lastZoneUpdate = time.Now()
-	}
-
-	reqZone := plugin.Zones(a.zones).Matches(reqName)
-	if reqZone == "" {
-		return plugin.NextOrFailure(a.Name(), a.Next, ctx, w, r)
-	}
-
-	// handle soa record
-	if reqType == dns.TypeSOA {
-		soaRec, err := a.getSOARecord(ctx, client, reqZone)
-		if err != nil {
-			return a.errorResponse(req, dns.RcodeServerFailure, err)
-		}
-		log.Info(soaRec)
-
-	}
-
-	// TODO:(jproxx) do something with with the rrs
-	rrs, err := a.getRRecords(ctx, client, reqName, reqType)
-	if err != nil {
-		return a.errorResponse(req, dns.RcodeServerFailure, err)
-	}
-	log.Info("rec ", len(rrs))
-
 	// Wrap.
-	pw := NewResponsePrinter(w)
+	// pw := NewResponsePrinter(w)
 
-	if len(rrs) == 0 {
-		// TODO: nothing found - check what we have to answer
-		// TODO: some plugins are sending SOA rr...
-		plugin.NextOrFailure(a.Name(), a.Next, ctx, pw, r)
-	}
+	responseMsg := new(dns.Msg)
+	responseMsg.SetReply(r)
+	responseMsg.Authoritative = true
+	responseMsg.RecursionAvailable = false
+	responseMsg.Compress = true
+
+	state := request.Request{W: w, Req: r}
+
+	reqName := state.Name()
+	reqType := state.QType()
 
 	// Export metric with the server label set to the current server handling the request.
 	requestCount.WithLabelValues(metrics.WithServer(ctx)).Inc()
 
-	// Call next plugin (if any).
-	return plugin.NextOrFailure(a.Name(), a.Next, ctx, pw, r)
+	client, err := a.getAtlasClient()
+	if err != nil {
+		return a.errorResponse(state, dns.RcodeServerFailure, err)
+	}
+	defer client.Close()
+
+	if a.mustReloadZones() {
+		err := a.loadZones(ctx, client)
+		if err != nil {
+			return a.errorResponse(state, dns.RcodeServerFailure, err)
+		}
+	}
+
+	reqZone := plugin.Zones(a.zones).Matches(reqName)
+	if reqZone == "" {
+		log.Errorf("zone not found: %v", state.Name())
+		return plugin.NextOrFailure(a.Name(), a.Next, ctx, w, r)
+	}
+	log.Infof("zone found: %v for question: %v", reqZone, reqName)
+
+	// TODO(jproxx): implement axfr requests, because we are needing them
+	if reqType == dns.TypeAXFR {
+		return a.errorResponse(state, dns.RcodeNotImplemented, nil)
+	}
+
+	// handle soa record
+	var soaRec []dns.RR
+	if reqType == dns.TypeSOA {
+		soaRec, err = a.getSOARecord(ctx, client, reqZone)
+		if err != nil {
+			return a.errorResponse(state, dns.RcodeServerFailure, err)
+		}
+		responseMsg.Answer = append(responseMsg.Answer, soaRec...)
+
+		state.SizeAndDo(responseMsg)
+		responseMsg = state.Scrub(responseMsg)
+		if err = w.WriteMsg(responseMsg); err != nil {
+			return dns.RcodeServerFailure, err
+		}
+
+		return dns.RcodeSuccess, nil
+	}
+
+	rrs, err := a.getRRecords(ctx, client, reqName, reqType)
+	if err != nil {
+		return a.errorResponse(state, dns.RcodeServerFailure, err)
+	}
+
+	if len(rrs) == 0 {
+		// to get bind compliance we are returning the SOA rr
+		soaRec, err = a.getSOARecord(ctx, client, reqZone)
+		if err != nil {
+			return a.errorResponse(state, dns.RcodeServerFailure, err)
+		}
+		responseMsg.Answer = append(responseMsg.Answer, soaRec...)
+
+		state.SizeAndDo(responseMsg)
+		responseMsg = state.Scrub(responseMsg)
+		if err = w.WriteMsg(responseMsg); err != nil {
+			log.Error(err)
+			return dns.RcodeServerFailure, err
+		}
+
+		return dns.RcodeSuccess, nil
+	}
+
+	responseMsg.Answer = append(responseMsg.Answer, rrs...)
+
+	// build authoritative section
+	if nsRRs, err := a.getNameServers(ctx, client, reqZone); err != nil {
+		return a.errorResponse(state, dns.RcodeServerFailure, err)
+	} else {
+		responseMsg.Ns = nsRRs
+	}
+
+	state.SizeAndDo(responseMsg)
+	responseMsg = state.Scrub(responseMsg)
+	if err = w.WriteMsg(responseMsg); err != nil {
+		log.Error(err)
+		return dns.RcodeServerFailure, err
+	}
+
+	return dns.RcodeSuccess, nil
 }
 
 // Name implements the Handler interface.
-func (a Atlas) Name() string { return plgName }
+func (a *Atlas) Name() string { return plgName }
 
-func (a Atlas) getAtlasClient() (client *ent.Client, err error) {
-	dsn, err := a.cfg.GetDsn()
-	if err != nil {
-		return nil, err
-	}
-
-	client, err = OpenAtlasDB(dsn)
-	if err != nil {
-		return nil, err
-	}
-
-	if a.cfg.debug {
-		client = client.Debug()
-	}
-	return
-}
-
-func (a *Atlas) loadZones(ctx context.Context, client *ent.Client) error {
-	zones := []string{}
-
-	if client == nil {
-		return fmt.Errorf("atlas client error")
-	}
-
-	records, err := client.DnsZone.
-		Query().
-		Select(dnszone.FieldName).
-		Where(dnszone.Activated(true)).
-		Order(ent.Asc(dnszone.FieldName)).
-		All(ctx)
-	if err != nil {
-		return err
-	}
-
-	log.Infof("loadZones found: %v zone(s)", len(records))
-
-	for _, zone := range records {
-		zones = append(zones, zone.Name)
-	}
-
-	a.zones = zones
-
-	return nil
-}
-
-func (a Atlas) getSOARecord(ctx context.Context, client *ent.Client, zone string) (rrs []dns.RR, err error) {
-	rrs = make([]dns.RR, 0)
-	if client == nil {
-		return rrs, fmt.Errorf("atlas client error")
-	}
-
-	soaRec, err := client.DnsZone.Query().
-		Where(
-			dnszone.Activated(true),
-			dnszone.NameEQ(zone),
-		).
-		First(ctx)
-
-	if err != nil {
-		return rrs, err
-	}
-
-	rec := &dns.SOA{
-		Hdr: dns.RR_Header{
-			Name:   soaRec.Name,
-			Rrtype: soaRec.Rrtype,
-			Class:  soaRec.Class,
-			Ttl:    soaRec.TTL,
-		},
-		Ns:      soaRec.Ns,
-		Mbox:    soaRec.Mbox,
-		Serial:  soaRec.Serial,
-		Refresh: soaRec.Refresh,
-		Retry:   soaRec.Retry,
-		Expire:  soaRec.Expire,
-		Minttl:  soaRec.Minttl,
-	}
-
-	return []dns.RR{rec}, nil
-}
-
-func (a Atlas) getRRecords(ctx context.Context, client *ent.Client, reqName string, reqQType uint16) (rrs []dns.RR, err error) {
-	rrs = make([]dns.RR, 0)
-	if client == nil {
-		return rrs, fmt.Errorf("atlas client error")
-	}
-
-	records, err := client.DnsRR.Query().
-		Select(
-			dnsrr.FieldName,
-			dnsrr.FieldClass,
-			dnsrr.FieldRrtype,
-			dnsrr.FieldRrdata,
-			dnsrr.FieldTTL,
-		).
-		Where(
-			dnsrr.NameEQ(reqName),
-			dnsrr.RrtypeEQ(reqQType),
-			dnsrr.ActivatedEQ(true), // we serve only activated records
-		).
-		Order(ent.Asc(
-			dnsrr.FieldName,
-			dnsrr.FieldRrtype,
-		)).
-		All(ctx)
-	if err != nil {
-		return rrs, err
-	}
-
-	for _, r := range records {
-		rec, err := record.From(r)
-		if err != nil {
-			log.Error(err)
-			return rrs, err
-		}
-		rrs = append(rrs, rec)
-	}
-
-	return rrs, nil
+func (a *Atlas) mustReloadZones() bool {
+	return time.Since(a.lastZoneUpdate) > a.cfg.zoneUpdateTime
 }
 
 func (handler *Atlas) errorResponse(state request.Request, rCode int, err error) (int, error) {
@@ -241,6 +159,5 @@ func NewResponsePrinter(w dns.ResponseWriter) *ResponsePrinter {
 
 // WriteMsg calls the underlying ResponseWriter's WriteMsg method and prints "atlas" to standard output.
 func (r *ResponsePrinter) WriteMsg(res *dns.Msg) error {
-	log.Info(plgName)
 	return r.ResponseWriter.WriteMsg(res)
 }
